Create a python file in prepare_dataset to automatically generate file list
